[Fix][GEMM] Tie the W4A16 weight scale to the activation dtype
`weight_scale` was declared FP32 while both reference kernels require it to
match the activation. Machete raises `prepack_B_dispatch NotImplementedError`
for any other dtype; Marlin does not check and returns `[inf, -1531.0, nan]`
against a reference the matching dtype reproduces exactly. An FP32 scale was
therefore not "also allowed" -- it was silently wrong on one baseline and
rejected by the other.

The manifest now says `same_as(activation)`, and the op's class docstring
states what it commits to: unconstrained N, K a multiple of `group_size`,
exact products in an FP32 accumulation, and what is explicitly out of scope.

+++ benchmarks/ops/bench_gemm.py
@@ -338,7 +338,7 @@ def _prepare_marlin_w4a16_baseline(
         get_weight_perm(4),
     )
     scales = marlin_permute_scales(
-        weight_scale.T.to(torch.float16).contiguous(),
+        weight_scale.T.to(activation.dtype).contiguous(),
         k,
         n,
         GROUP_SIZE,
